% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bar_theme}
\alias{bar_theme}
\title{Bar plot themes}
\usage{
bar_theme(base_size = 10, orientation = c("vertical", "horizontal"))
}
\arguments{
\item{base_size}{Sent to \code{\link{theme_bw}}.}

\item{orientation}{Is it a vertical or horizontal bar plot?}
}
\description{
Bar plot themes
}
\examples{
library(ggplot2)

plot_data <- data.frame(
  category = LETTERS[1:5],
  value = c(4, 5, 2, 3, 2)
)
p <- ggplot(plot_data, aes(x = category, y = value)) +
  geom_bar(stat = "identity") +
  scale_y_continuous(expand = c(0, 0))
p + bar_theme(16)
p + coord_flip() + bar_theme(16, orientation="horizontal")

}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{line_theme}}.
}

