% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{colour}
\alias{colour}
\alias{default_palette}
\alias{get_palette}
\alias{validate_palette}
\alias{set_palette}
\title{Colours and palettes}
\format{An object of class \code{character} of length 5.}
\usage{
colour(name, palette = get_palette())

default_palette

get_palette()

validate_palette(palette)

set_palette(palette)
}
\arguments{
\item{name}{Name of colour.}

\item{palette}{Named character vector. See}
}
\description{
Colours and palettes
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\keyword{datasets}
