% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{emphasis_colour}
\alias{emphasis_colour}
\alias{faint_colour}
\alias{light_colour}
\alias{main_colour}
\alias{pale_colour}
\title{Colours used by trees, maps, and theorems}
\format{An object of class \code{character} of length 1.}
\usage{
emphasis_colour

main_colour

pale_colour

light_colour

faint_colour
}
\description{
Colours used by trees, maps, and theorems
}
\keyword{datasets}

