% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{numeric_label}
\alias{numeric_label}
\title{Numeric labels for internal use and special cases}
\usage{
numeric_label(x, prefix = NULL, unit = "", unit_position = c("after",
  "before"), digits = 3, space = TRUE)
}
\arguments{
\item{x}{Numeric values.}

\item{prefix}{A data frame containing prefix symbols and powers,
see \link{prefix}.}

\item{unit}{Unit symbol.}

\item{unit_position}{Whether to put the unit before or after the numerical
value.}

\item{digits}{Number of significant digits, sent to \code{\link{signif}}.}

\item{space}{Whether to put a space between numeric value and unit.}
}
\value{
Character vector.
}
\description{
Low level function for generating all other types of numeric labels.
The end-user should primarily use functions listed in
\code{\link{quantity_label}}, whereas this function is intended for internal
use and special cases. If you find that you are
regularly using \code{numeric_label} for a case not covered by the end-user
functions, please file an issue on GitHub for adding it.
}
\examples{
# Physical quantities
n <- 12
x <- 10^rnorm(n, sd = 6) * ifelse(runif(n) < .5, -1, 1)

data.frame(
  x = x,
  label = quantity_label(x, unit = "g")
)

# Currencies
x <- round(10^c(rgamma(n/2, 3, 1) - 2, rgamma(n/2, 5, 1)), digits = 2) * ifelse(runif(n) < .5, -1, 1)

data.frame(
  x = x,
  EUR = currency_label(x, prefix = TRUE, currency = "EUR"),
  USD = currency_label(x, prefix = TRUE, currency = "USD"),
  GBP = currency_label(x, prefix = TRUE, currency = "GBP"),
  SEK = currency_label(x, prefix = TRUE, currency = "SEK")
)
}
\seealso{
\code{\link{quantity_label}}, \code{\link{prefix}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
