% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{quantity_label}
\alias{quantity_label}
\alias{currency_label}
\alias{percentage_label}
\alias{logical_label}
\title{Labels for common cases}
\usage{
quantity_label(x, prefix = TRUE, unit = "", ...)

currency_label(x, prefix = TRUE, currency = "EUR", ...)

percentage_label(x, unit = "\%", ...)

logical_label(x, levels = c(FALSE, NA, TRUE))
}
\arguments{
\item{x}{Numeric values.}

\item{prefix}{Whether to use unit prefix, e.g. "k" for "kilo".}

\item{unit}{Unit of the quantity, e.g. "m" for "meter".}

\item{...}{Sent to the low level function \code{\link{numeric_label}}.}

\item{currency}{Currency abbreviation. So far only EUR, GBP, USD, and SEK are
supported.}

\item{levels}{Which level to keep. Note that the order determines the factor
level order.}
}
\value{
Factor.
}
\description{
Labels for common cases

Logical labels
}
\seealso{
\code{\link{numeric_label}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
