% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{shift_column}
\alias{shift_column}
\title{Groupwise shift the values of a column to make their ranges non-overlapping}
\usage{
shift_column(.data, value_col, facet_col, fun = function(v) range(v, na.rm =
  TRUE))
}
\arguments{
\item{.data}{Data frame.}

\item{value_col}{Name of the column to shift.}

\item{facet_col}{Name of the column that will later be used for faceting.}

\item{fun}{Function that produce axis breaks. If you wish to add more breaks
please do so by manipulating the return value (2nd element).}
}
\value{
A list of two elements:
\itemize{
	 \item{A data frame with shifted data.}
	 \item{A data frame with the mapping between original and shifted breaks.}
}
}
\description{
Intended to be used as a hack to get different y-axes breaks.
}
\examples{
library(ggplot2)

break_fun <- function(v){
	c(mean(v), range(v))
}

shifted_cars <- shift_column(mtcars, "mpg", facet_col = "cyl", fun = break_fun)

plot_data <- shifted_cars[[1]]
mapping <- shifted_cars[[2]]

ggplot(plot_data, aes(x = disp, y = mpg)) +
  facet_wrap(~cyl, scales = "free_y") +
  geom_point() +
  scale_y_continuous(
    breaks = mapping$breaks,
    labels = numeric_label(mapping$labels),
    expand = c(0, 0)
  ) +
	line_theme(box = TRUE, grid = "y")

}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
