% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{temp_column_name}
\alias{temp_column_name}
\title{Generate a temporary column name}
\usage{
temp_column_name(.data)
}
\arguments{
\item{.data}{Data frame to match names against, to ensure uniqueness.}
}
\value{
A random string not present in \code{names(.data)}.
}
\description{
Generate a temporary column name
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
