% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{tmt_label}
\alias{tmt_label}
\title{Geoms with TMT defaults}
\usage{
tmt_label(..., colour = colour("annotation_light"))
}
\arguments{
\item{...}{Sent to \pkg{ggplot2} geoms.}

\item{colour}{Colour.}
}
\description{
Geoms with TMT defaults
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
