% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tmt_save}
\alias{tmt_save}
\title{Save a plot}
\usage{
tmt_save(filename, plot, width, height, units = "cm", ...)
}
\arguments{
\item{filename}{Output filename.}

\item{plot}{Plot drawn with ggplot.}

\item{width}{Output width (in cm unless \code{units} is set).}

\item{height}{Output height (in cm unless \code{units} is set).}

\item{units}{Unit for \code{width} and \code{height}.}

\item{...}{Sent to \code{\link{ggsave}}.}
}
\description{
Sets a few default values for \code{\link{ggsave}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
