% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{tmt_theme}
\alias{tmt_theme}
\title{Create trees-maps-and-theorems theme}
\usage{
tmt_theme(base_size = 10, axis.line = "", axis.text = "",
  axis.ticks = "", axis.title = "", panel.border = FALSE,
  panel.grid = "")
}
\arguments{
\item{base_size}{Sent to \code{\link{theme_bw}}.}

\item{axis.line}{Axis lines (xy-string).}

\item{axis.text}{Axis tick labels (xy-string).}

\item{axis.ticks}{Axis tick lines (xy-string).}

\item{axis.title}{Axis titles (xy-string).}

\item{panel.border}{Box surrounding each panel (logical).}

\item{panel.grid}{Grid lines (xy-string). Only major grid lines are ever used.}
}
\description{
This function provides the lowest level entry point to the tmt-engine
intended for the end user.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
\code{\link{bar_theme}}, \code{\link{line_theme}}.
}

